requirements.txt
bidict==0.22.1
blinker==1.7.0
boto3==1.33.8
botocore==1.33.8
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.8.2
simple-websocket==1.0.0
six==1.16.0
socketIO-client==0.7.2
urllib3==2.0.7
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
python-dotenv
boto3
requests
flask